Solves #497
